+++ src/app_context.rs
@@ -11,7 +11,7 @@ use crate::data_loader::DataLoader;
 use crate::graphql::GraphqlDataLoader;
 use crate::grpc::data_loader::GrpcDataLoader;
 use crate::http::{DataLoaderRequest, HttpDataLoader};
-use crate::lambda::{DataLoaderId, Expression, Unsafe};
+use crate::lambda::{DataLoaderId, Expression, Io};
 use crate::{grpc, EnvIO, HttpIO};
 
 pub struct AppContext<Http, Env> {
@@ -36,9 +36,9 @@ impl<Http: HttpIO, Env: EnvIO> AppContext<Http, Env> {
     for def in blueprint.definitions.iter_mut() {
       if let Definition::ObjectTypeDefinition(def) = def {
         for field in &mut def.fields {
-          if let Some(Expression::Unsafe(expr_unsafe)) = &mut field.resolver {
+          if let Some(Expression::Io(expr_unsafe)) = &mut field.resolver {
             match expr_unsafe {
-              Unsafe::Http { req_template, group_by, .. } => {
+              Io::Http { req_template, group_by, .. } => {
                 let data_loader = HttpDataLoader::new(
                   h_client.clone(),
                   group_by.clone(),
@@ -46,7 +46,7 @@ impl<Http: HttpIO, Env: EnvIO> AppContext<Http, Env> {
                 )
                 .to_data_loader(blueprint.upstream.batch.clone().unwrap_or_default());
 
-                field.resolver = Some(Expression::Unsafe(Unsafe::Http {
+                field.resolver = Some(Expression::Io(Io::Http {
                   req_template: req_template.clone(),
                   group_by: group_by.clone(),
                   dl_id: Some(DataLoaderId(http_data_loaders.len())),
@@ -55,11 +55,11 @@ impl<Http: HttpIO, Env: EnvIO> AppContext<Http, Env> {
                 http_data_loaders.push(data_loader);
               }
 
-              Unsafe::GraphQLEndpoint { req_template, field_name, batch, .. } => {
+              Io::GraphQLEndpoint { req_template, field_name, batch, .. } => {
                 let graphql_data_loader = GraphqlDataLoader::new(h_client.clone(), *batch)
                   .to_data_loader(blueprint.upstream.batch.clone().unwrap_or_default());
 
-                field.resolver = Some(Expression::Unsafe(Unsafe::GraphQLEndpoint {
+                field.resolver = Some(Expression::Io(Io::GraphQLEndpoint {
                   req_template: req_template.clone(),
                   field_name: field_name.clone(),
                   batch: *batch,
@@ -69,7 +69,7 @@ impl<Http: HttpIO, Env: EnvIO> AppContext<Http, Env> {
                 gql_data_loaders.push(graphql_data_loader);
               }
 
-              Unsafe::Grpc { req_template, group_by, .. } => {
+              Io::Grpc { req_template, group_by, .. } => {
                 let data_loader = GrpcDataLoader {
                   client: h2_client.clone(),
                   operation: req_template.operation.clone(),
@@ -77,7 +77,7 @@ impl<Http: HttpIO, Env: EnvIO> AppContext<Http, Env> {
                 };
                 let data_loader = data_loader.to_data_loader(blueprint.upstream.batch.clone().unwrap_or_default());
 
-                field.resolver = Some(Expression::Unsafe(Unsafe::Grpc {
+                field.resolver = Some(Expression::Io(Io::Grpc {
                   req_template: req_template.clone(),
                   group_by: group_by.clone(),
                   dl_id: Some(DataLoaderId(grpc_data_loaders.len())),
