+++ src/blueprint/validate.rs
@@ -1,6 +1,6 @@
 use super::{is_scalar, to_type, FieldDefinition, Type};
 use crate::config::{self, Config};
-use crate::lambda::{Expression, Unsafe};
+use crate::lambda::{Expression, Io};
 use crate::valid::Valid;
 
 struct MustachePartsValidator<'a> {
@@ -111,7 +111,7 @@ impl FieldDefinition {
     let parts_validator = MustachePartsValidator::new(type_of, config, self);
 
     match &self.resolver {
-      Some(Expression::Unsafe(Unsafe::Http { req_template, .. })) => {
+      Some(Expression::Io(Io::Http { req_template, .. })) => {
         Valid::from_iter(req_template.root_url.expression_segments(), |parts| {
           parts_validator.validate(parts, false).trace("path")
         })
@@ -124,7 +124,7 @@ impl FieldDefinition {
         }))
         .unit()
       }
-      Some(Expression::Unsafe(Unsafe::GraphQLEndpoint { req_template, .. })) => {
+      Some(Expression::Io(Io::GraphQLEndpoint { req_template, .. })) => {
         Valid::from_iter(req_template.headers.clone(), |(_, mustache)| {
           Valid::from_iter(mustache.expression_segments(), |parts| {
             parts_validator.validate(parts, true).trace("headers")
@@ -143,7 +143,7 @@ impl FieldDefinition {
         })
         .unit()
       }
-      Some(Expression::Unsafe(Unsafe::Grpc { req_template, .. })) => {
+      Some(Expression::Io(Io::Grpc { req_template, .. })) => {
         Valid::from_iter(req_template.url.expression_segments(), |parts| {
           parts_validator.validate(parts, false).trace("path")
         })
