+++ src/lambda/expression.rs
@@ -1,36 +1,21 @@
 use std::fmt::Debug;
-use std::sync::Arc;
 
 use anyhow::Result;
 use async_graphql_value::ConstValue;
-use reqwest::Request;
 use serde_json::Value;
 use thiserror::Error;
 
 use super::list::List;
 use super::logic::Logic;
-use super::{Eval, Math, Relation, ResolverContextLike};
-use crate::config::group_by::GroupBy;
-use crate::config::GraphQLOperationType;
-use crate::data_loader::{DataLoader, Loader};
-use crate::graphql::{self, GraphqlDataLoader};
-use crate::grpc;
-use crate::grpc::data_loader::GrpcDataLoader;
-use crate::grpc::protobuf::ProtobufOperation;
-use crate::grpc::request::execute_grpc_request;
-use crate::grpc::request_template::RenderedRequestTemplate;
-use crate::http::{self, cache_policy, DataLoaderRequest, HttpDataLoader, Response};
-#[cfg(feature = "unsafe-js")]
-use crate::javascript;
+use super::{Eval, EvaluationContext, Io, Math, Relation, ResolverContextLike};
 use crate::json::JsonLike;
-use crate::lambda::EvaluationContext;
 
 #[derive(Clone, Debug)]
 pub enum Expression {
   Context(Context),
   Literal(Value), // TODO: this should async_graphql::Value
   EqualTo(Box<Expression>, Box<Expression>),
-  Unsafe(Unsafe),
+  Io(Io),
   Input(Box<Expression>, Vec<String>),
   Logic(Logic),
   Relation(Relation),
@@ -51,30 +36,6 @@ pub enum Context {
   Path(Vec<String>),
 }
 
-#[derive(Clone, Debug)]
-pub enum Unsafe {
-  Http {
-    req_template: http::RequestTemplate,
-    group_by: Option<GroupBy>,
-    dl_id: Option<DataLoaderId>,
-  },
-  GraphQLEndpoint {
-    req_template: graphql::RequestTemplate,
-    field_name: String,
-    batch: bool,
-    dl_id: Option<DataLoaderId>,
-  },
-  Grpc {
-    req_template: grpc::RequestTemplate,
-    group_by: Option<GroupBy>,
-    dl_id: Option<DataLoaderId>,
-  },
-  JS(Box<Expression>, String),
-}
-
-#[derive(Clone, Copy, Debug)]
-pub struct DataLoaderId(pub usize);
-
 #[derive(Debug, Error)]
 pub enum EvaluationError {
   #[error("IOException: {0}")]
@@ -86,14 +47,8 @@ pub enum EvaluationError {
   #[error("APIValidationError: {0:?}")]
   APIValidationError(Vec<String>),
 
-  #[error("ConcatException: {0:?}")]
-  ConcatException(String),
-
-  #[error("IntersectionException: {0:?}")]
-  IntersectionException(String),
-
-  #[error("OperationFailed: {0:?}")]
-  OperationFailed(String),
+  #[error("ExprEvalError: {0:?}")]
+  ExprEvalError(String),
 }
 
 impl<'a> From<crate::valid::ValidationError<&'a str>> for EvaluationError {
@@ -144,85 +99,7 @@ impl Eval for Expression {
       Expression::EqualTo(left, right) => Ok(async_graphql::Value::from(
         left.eval(ctx, conc).await? == right.eval(ctx, conc).await?,
       )),
-      Expression::Unsafe(operation) => match operation {
-        Unsafe::Http { req_template, dl_id, .. } => {
-          let req = req_template.to_request(ctx)?;
-          let is_get = req.method() == reqwest::Method::GET;
-
-          let res = if is_get && ctx.req_ctx.is_batching_enabled() {
-            let data_loader: Option<&DataLoader<DataLoaderRequest, HttpDataLoader>> =
-              dl_id.and_then(|index| ctx.req_ctx.http_data_loaders.get(index.0));
-            execute_request_with_dl(ctx, req, data_loader).await?
-          } else {
-            execute_raw_request(ctx, req).await?
-          };
-
-          if ctx.req_ctx.server.get_enable_http_validation() {
-            req_template
-              .endpoint
-              .output
-              .validate(&res.body)
-              .to_result()
-              .map_err(EvaluationError::from)?;
-          }
-
-          set_cache_control(ctx, &res);
-
-          Ok(res.body)
-        }
-        Unsafe::GraphQLEndpoint { req_template, field_name, dl_id, .. } => {
-          let req = req_template.to_request(ctx)?;
-
-          let res = if ctx.req_ctx.upstream.batch.is_some()
-            && matches!(req_template.operation_type, GraphQLOperationType::Query)
-          {
-            let data_loader: Option<&DataLoader<DataLoaderRequest, GraphqlDataLoader>> =
-              dl_id.and_then(|index| ctx.req_ctx.gql_data_loaders.get(index.0));
-            execute_request_with_dl(ctx, req, data_loader).await?
-          } else {
-            execute_raw_request(ctx, req).await?
-          };
-
-          set_cache_control(ctx, &res);
-          parse_graphql_response(ctx, res, field_name)
-        }
-        Unsafe::Grpc { req_template, dl_id, .. } => {
-          let rendered = req_template.render(ctx)?;
-
-          let res = if ctx.req_ctx.upstream.batch.is_some() &&
-                // TODO: share check for operation_type for resolvers
-                matches!(req_template.operation_type, GraphQLOperationType::Query)
-          {
-            let data_loader: Option<&DataLoader<grpc::DataLoaderRequest, GrpcDataLoader>> =
-              dl_id.and_then(|index| ctx.req_ctx.grpc_data_loaders.get(index.0));
-            execute_grpc_request_with_dl(ctx, rendered, data_loader).await?
-          } else {
-            let req = rendered.to_request()?;
-            execute_raw_grpc_request(ctx, req, &req_template.operation).await?
-          };
-
-          set_cache_control(ctx, &res);
-
-          Ok(res.body)
-        }
-        Unsafe::JS(input, script) => {
-          let result;
-          #[cfg(not(feature = "unsafe-js"))]
-          {
-            let _ = script;
-            let _ = input;
-            result = Err(EvaluationError::JSException("JS execution is disabled".to_string()).into());
-          }
-
-          #[cfg(feature = "unsafe-js")]
-          {
-            let input = input.eval(ctx, conc).await?;
-            result = javascript::execute_js(script, input, Some(ctx.timeout))
-              .map_err(|e| EvaluationError::JSException(e.to_string()).into());
-          }
-          result
-        }
-      },
+      Expression::Io(operation) => operation.async_eval(ctx, conc).await,
 
       Expression::Relation(relation) => relation.async_eval(ctx, conc).await,
       Expression::Logic(logic) => logic.async_eval(ctx, conc).await,
@@ -255,112 +132,6 @@ pub fn get_path_for_const_value_ref<'a>(
   }
 
   Some(const_value)
-}
-
-fn set_cache_control<'ctx, Ctx: ResolverContextLike<'ctx>>(
-  ctx: &EvaluationContext<'ctx, Ctx>,
-  res: &Response<async_graphql::Value>,
-) {
-  if ctx.req_ctx.server.get_enable_cache_control() && res.status.is_success() {
-    if let Some(policy) = cache_policy(res) {
-      ctx.req_ctx.set_cache_control(policy);
-    }
-  }
-}
-
-async fn execute_raw_request<'ctx, Ctx: ResolverContextLike<'ctx>>(
-  ctx: &EvaluationContext<'ctx, Ctx>,
-  req: Request,
-) -> Result<Response<async_graphql::Value>> {
-  ctx
-    .req_ctx
-    .h_client
-    .execute(req)
-    .await
-    .map_err(|e| EvaluationError::IOException(e.to_string()))?
-    .to_json()
-}
-
-async fn execute_raw_grpc_request<'ctx, Ctx: ResolverContextLike<'ctx>>(
-  ctx: &EvaluationContext<'ctx, Ctx>,
-  req: Request,
-  operation: &ProtobufOperation,
-) -> Result<Response<async_graphql::Value>> {
-  Ok(
-    execute_grpc_request(&ctx.req_ctx.h2_client, operation, req)
-      .await
-      .map_err(|e| EvaluationError::IOException(e.to_string()))?,
-  )
-}
-
-async fn execute_grpc_request_with_dl<
-  'ctx,
-  Ctx: ResolverContextLike<'ctx>,
-  Dl: Loader<grpc::DataLoaderRequest, Value = Response<async_graphql::Value>, Error = Arc<anyhow::Error>>,
->(
-  ctx: &EvaluationContext<'ctx, Ctx>,
-  rendered: RenderedRequestTemplate,
-  data_loader: Option<&DataLoader<grpc::DataLoaderRequest, Dl>>,
-) -> Result<Response<async_graphql::Value>> {
-  let headers = ctx
-    .req_ctx
-    .upstream
-    .batch
-    .clone()
-    .map(|s| s.headers)
-    .unwrap_or_default();
-  let endpoint_key = grpc::DataLoaderRequest::new(rendered, headers);
-
-  Ok(
-    data_loader
-      .unwrap()
-      .load_one(endpoint_key)
-      .await
-      .map_err(|e| EvaluationError::IOException(e.to_string()))?
-      .unwrap_or_default(),
-  )
-}
-
-async fn execute_request_with_dl<
-  'ctx,
-  Ctx: ResolverContextLike<'ctx>,
-  Dl: Loader<DataLoaderRequest, Value = Response<async_graphql::Value>, Error = Arc<anyhow::Error>>,
->(
-  ctx: &EvaluationContext<'ctx, Ctx>,
-  req: Request,
-  data_loader: Option<&DataLoader<DataLoaderRequest, Dl>>,
-) -> Result<Response<async_graphql::Value>> {
-  let headers = ctx
-    .req_ctx
-    .upstream
-    .batch
-    .clone()
-    .map(|s| s.headers)
-    .unwrap_or_default();
-  let endpoint_key = crate::http::DataLoaderRequest::new(req, headers);
-
-  Ok(
-    data_loader
-      .unwrap()
-      .load_one(endpoint_key)
-      .await
-      .map_err(|e| EvaluationError::IOException(e.to_string()))?
-      .unwrap_or_default(),
-  )
-}
-
-fn parse_graphql_response<'ctx, Ctx: ResolverContextLike<'ctx>>(
-  ctx: &EvaluationContext<'ctx, Ctx>,
-  res: Response<async_graphql::Value>,
-  field_name: &str,
-) -> Result<async_graphql::Value> {
-  let res: async_graphql::Response = serde_json::from_value(res.body.into_json()?)?;
-
-  for error in res.errors {
-    ctx.add_error(error);
-  }
-
-  Ok(res.data.get_key(field_name).map(|v| v.to_owned()).unwrap_or_default())
 }
 
 #[cfg(test)]
