+++ src/lambda/math.rs
@@ -33,7 +33,7 @@ impl Eval for Math {
 
         try_i64_operation(&lhs, &rhs, ops::Rem::rem)
           .or_else(|| try_u64_operation(&lhs, &rhs, ops::Rem::rem))
-          .ok_or(EvaluationError::OperationFailed("mod".into()))?
+          .ok_or(EvaluationError::ExprEvalError("mod".into()))?
       }
       Math::Add(lhs, rhs) => {
         let lhs = lhs.eval(ctx, conc).await?;
@@ -42,7 +42,7 @@ impl Eval for Math {
         try_f64_operation(&lhs, &rhs, ops::Add::add)
           .or_else(|| try_u64_operation(&lhs, &rhs, ops::Add::add))
           .or_else(|| try_i64_operation(&lhs, &rhs, ops::Add::add))
-          .ok_or(EvaluationError::OperationFailed("add".into()))?
+          .ok_or(EvaluationError::ExprEvalError("add".into()))?
       }
       Math::Dec(val) => {
         let val = val.eval(ctx, conc).await?;
@@ -53,7 +53,7 @@ impl Eval for Math {
           .map(|val| (val - 1f64).into())
           .or_else(|| val.as_u64_ok().ok().map(|val| (val - 1u64).into()))
           .or_else(|| val.as_i64_ok().ok().map(|val| (val - 1i64).into()))
-          .ok_or(EvaluationError::OperationFailed("dec".into()))?
+          .ok_or(EvaluationError::ExprEvalError("dec".into()))?
       }
       Math::Divide(lhs, rhs) => {
         let lhs = lhs.eval(ctx, conc).await?;
@@ -62,7 +62,7 @@ impl Eval for Math {
         try_f64_operation(&lhs, &rhs, ops::Div::div)
           .or_else(|| try_u64_operation(&lhs, &rhs, ops::Div::div))
           .or_else(|| try_i64_operation(&lhs, &rhs, ops::Div::div))
-          .ok_or(EvaluationError::OperationFailed("divide".into()))?
+          .ok_or(EvaluationError::ExprEvalError("divide".into()))?
       }
       Math::Inc(val) => {
         let val = val.eval(ctx, conc).await?;
@@ -73,7 +73,7 @@ impl Eval for Math {
           .map(|val| (val + 1f64).into())
           .or_else(|| val.as_u64_ok().ok().map(|val| (val + 1u64).into()))
           .or_else(|| val.as_i64_ok().ok().map(|val| (val + 1i64).into()))
-          .ok_or(EvaluationError::OperationFailed("dec".into()))?
+          .ok_or(EvaluationError::ExprEvalError("dec".into()))?
       }
       Math::Multiply(lhs, rhs) => {
         let lhs = lhs.eval(ctx, conc).await?;
@@ -82,7 +82,7 @@ impl Eval for Math {
         try_f64_operation(&lhs, &rhs, ops::Mul::mul)
           .or_else(|| try_u64_operation(&lhs, &rhs, ops::Mul::mul))
           .or_else(|| try_i64_operation(&lhs, &rhs, ops::Mul::mul))
-          .ok_or(EvaluationError::OperationFailed("multiply".into()))?
+          .ok_or(EvaluationError::ExprEvalError("multiply".into()))?
       }
       Math::Negate(val) => {
         let val = val.eval(ctx, conc).await?;
@@ -92,7 +92,7 @@ impl Eval for Math {
           .ok()
           .map(|val| (-val).into())
           .or_else(|| val.as_i64_ok().ok().map(|val| (-val).into()))
-          .ok_or(EvaluationError::OperationFailed("neg".into()))?
+          .ok_or(EvaluationError::ExprEvalError("neg".into()))?
       }
       Math::Product(exprs) => {
         let results: Vec<_> = exprs.eval(ctx, conc).await?;
@@ -101,7 +101,7 @@ impl Eval for Math {
           try_f64_operation(&lhs, &rhs, ops::Mul::mul)
             .or_else(|| try_u64_operation(&lhs, &rhs, ops::Mul::mul))
             .or_else(|| try_i64_operation(&lhs, &rhs, ops::Mul::mul))
-            .ok_or(EvaluationError::OperationFailed("product".into()))
+            .ok_or(EvaluationError::ExprEvalError("product".into()))
         })?
       }
       Math::Subtract(lhs, rhs) => {
@@ -111,7 +111,7 @@ impl Eval for Math {
         try_f64_operation(&lhs, &rhs, ops::Sub::sub)
           .or_else(|| try_u64_operation(&lhs, &rhs, ops::Sub::sub))
           .or_else(|| try_i64_operation(&lhs, &rhs, ops::Sub::sub))
-          .ok_or(EvaluationError::OperationFailed("subtract".into()))?
+          .ok_or(EvaluationError::ExprEvalError("subtract".into()))?
       }
       Math::Sum(exprs) => {
         let results: Vec<_> = exprs.eval(ctx, conc).await?;
@@ -120,7 +120,7 @@ impl Eval for Math {
           try_f64_operation(&lhs, &rhs, ops::Add::add)
             .or_else(|| try_u64_operation(&lhs, &rhs, ops::Add::add))
             .or_else(|| try_i64_operation(&lhs, &rhs, ops::Add::add))
-            .ok_or(EvaluationError::OperationFailed("sum".into()))
+            .ok_or(EvaluationError::ExprEvalError("sum".into()))
         })?
       }
     })
