+++ src/lambda/mod.rs
@@ -3,6 +3,7 @@ mod evaluation_context;
 mod expression;
 mod graphql_operation_context;
 mod helpers;
+mod io;
 mod lambda;
 mod list;
 mod logic;
@@ -15,6 +16,7 @@ pub use evaluation_context::EvaluationContext;
 pub(crate) use expression::*;
 pub use graphql_operation_context::GraphQLOperationContext;
 pub use helpers::*;
+pub use io::*;
 pub use lambda::Lambda;
 pub use list::*;
 pub use logic::*;
