+++ tests/graphql/test-unsafe.graphql
@@ -4,7 +4,7 @@ schema @server @upstream(baseURL: "http://jsonplacheholder.typicode.com") {
 }
 
 type Query {
-  foo: [User] @unsafe(script: "return [{id: 1, name: 'foo'}]")
+  foo: [User] @io(script: "return [{id: 1, name: 'foo'}]")
 }
 
 type User {
